+++ packages/runtime/tce-edit-vue3/src/main.ts
@@ -1,19 +1,32 @@
 import { createApp } from 'vue';
 
 import App from './App.vue';
+import ContentElement from './components/ContentElement.vue';
 import EmbeddedContainer from './components/EmbeddedContainer.vue';
 import NotCompositeAlert from './components/NotCompositeAlert.vue';
 import Radio from './radio';
 import vuetify from './plugins/vuetify';
 
 const element = await import(import.meta.env.EDIT_DIR);
-const { isComposite = false, isQuestion, isGradable } = element.default;
+const {
+  isComposite = false,
+  isQuestion,
+  isGradable,
+  name,
+  ui,
+} = element.default;
 
-const app = createApp(App, { isQuestion, isGradable });
+const app = createApp(App, {
+  isQuestion,
+  isGradable,
+  type: name,
+  icon: ui.icon,
+});
 const radio = Radio.getInstance();
 app.provide('$eventBus', radio);
 app.provide('$elementBus', radio.channel('app'));
 app.use(vuetify);
+app.component('TailorContentElement', ContentElement);
 app.component(
   'TailorEmbeddedContainer',
   isComposite ? EmbeddedContainer : NotCompositeAlert,