+++ packages/runtime/tce-edit-vue3/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tailor-cms/tce-edit-next-runtime
 
+## 0.6.32
+
+### Patch Changes
+
+- Fixed embed element delete.
+
+## 0.6.31
+
+### Patch Changes
+
+- Updated lodash imports.
+
+## 0.6.30
+
+### Patch Changes
+
+- Tweaked embedded container, extracted ContentElement component and wrapped questions in Question Card.
+
 ## 0.6.29
 
 ### Patch Changes